+++ spec/pgchief/config/s3_spec.rb
@@ -22,4 +22,13 @@ RSpec.describe Pgchief::Config::S3 do
 
     expect(s3_config.path).to eq("path/with-dashes/")
   end
+
+  it "is backwards compatible with s3_path_prefix" do
+    toml_file = File.expand_path("spec/fixtures/config_with_old_path_config.toml")
+    Pgchief::Config.load_config!(toml_file)
+
+    s3_config = described_class.new(Pgchief::Config)
+
+    expect(s3_config.path).to eq("path/")
+  end
 end